# Persist engine events in batches