Update: The required service-side changes have been rolled out. This is ready to review now.

---

When persisting engine events, send them in large batches rather than individually. This not only leads to a noticeable increase in "throughput", but also can reduce the variability in response time.

Currently, we persist engine events one-by-one, concurrently. Spawning a new go-routine to make the HTTP request to the Pulumi Service every time an event is encountered. When we first landed, https://github.com/pulumi/pulumi/pull/2314, it was clear this approach would lead to problems elsewhere but it actually worked surprisingly well the past six months.

However, over time this has has lead to some problems. For overly chatty updates, e.g. those spawning tens of thousands or hundreds of thousands of engine events, this would lead to several types of errors:

- I haven't tried to prove it just yet, but I'm pretty sure that it's possible for an update to fail because the go-routine that updates the lease token gets starved (because the system is busy dealing with all of the concurrent API requests).
- The Pulumi Service's response latency for engine events is very low, except occasionally some requests from clients will take 1-2 minutes. It seems that the problem is for whatever reason the HTTP request is started on the machine, and just gets starved -- despite other engine events being sent from the same machine. I haven't tried to prove this either, but I would guess it is possible that if the NIC or OS is overwhelmed with HTTP requests, it's possible to get into this state.

Sending engine events in batches addresses this by dramatically cutting down on concurrent network requests. The tradeoff is that the "payloads" of these requests can be 50x larger, but the majority of engine events are actually pretty small. (Since they are JSON serialized and then gzip-compressed before being sent.)

There are two key changes with this PR:

- First we buffer the channel sending engine events off for persistence (and to be displayed in the terminal). So if persistence "gets backed up" for some reason, there is a little wiggle room for it to catch up.
- We use a second, unbuffered channel specifically for persisting engine events. This way we can cap the number of concurrent HTTP requests to the Pulumi Service, so that sending many engine events wouldn't overwhelm the threadpool and/or network bandwidth. (It does mean however that if the CLI cannot transmit engine events fast enough, it would eventually cause the update to be blocked until they could be sent off.)

---

I won't have perf numbers for what to expect until I can deploy the service-side changes to production. But here are some stats from running the current `ee_perf` test against my devstack (now located in Mumbai (AWS `ap-southeast-1`)

The `ee_perf` test just sends 1,000 `DiagEvents` in a tight for loop. In December, that took ~4s, and has gotten on average a little slower and more unpredictable when ran from Travis CI jobs. (Sometimes almost taking a minute to complete!)

**Current, unlimited go-routines approach**
Times to run the test on my laptop, connecting to a devstack halfway-across the world. (So I would expect US-US traffic to be much faster.)

12.32s
11.56s
11.40s
10.20s
9.32s
16.24s

4x attempts didn't even complete, because my dev stack was overwhelmed and returned a 500 or 502.

**With the batched approach**
With the patch applied, we are back under that 8s limit, even though the devstack is in Mumbai. When rolled out to production (with its much larger database capacity), and larger number of service workers, I would think these numbers would be slightly better.

5.68s
5.33s
5.22s
6.37s

When connecting to the production instance on the west-coast, the numbers are even better:

2.22s
1.79s
1.79s
2.09s