@chrsmith 2019-06-25
@ellismg , @pgavlin the service-side changes have rolled out to production and so you can run/test this without needing to change the Pulumi API URL. Please take a look when convenient.